// packages/oos-dsls-ts/src/generated/ast.ts
export interface Accordion extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Accordion';
    items: Array<AccordionItem>;
    mods: Array<LayoutMod>;
}

export const Accordion = {
    $type: 'Accordion',
    items: 'items',
    mods: 'mods'
} as const;

export function isAccordion(item: unknown): item is Accordion {
    return reflection.isInstance(item, Accordion.$type);
}

export interface AccordionItem extends langium.AstNode {
    readonly $container: Accordion;
    readonly $type: 'AccordionItem';
    body: Array<BodyElement>;
    caption: string;
    open: boolean;
}

export const AccordionItem = {
    $type: 'AccordionItem',
    body: 'body',
    caption: 'caption',
    open: 'open'
} as const;
// ...
export interface As extends langium.AstNode {
    readonly $container: ToolbarNew | ViewAction;
    readonly $type: 'As';
    target: PresentationTarget;
}

export const As = {
    $type: 'As',
    target: 'target'
} as const;
// ...
export interface Bind extends langium.AstNode {
    readonly $container: ToolbarNew | ViewAction;
    readonly $type: 'Bind';
    source: FieldRef;
}

export const Bind = {
    $type: 'Bind',
    source: 'source'
} as const;

export function isBind(item: unknown): item is Bind {
    return reflection.isInstance(item, Bind.$type);
}

export type BodyElement = Accordion | Button | Card | Divider | Grid | IconWidget | LinkWidget | RichText | Row | Section | Separator | Stack | Table | Tabs | Widget;

export const BodyElement = {
    $type: 'BodyElement'
} as const;

export function isBodyElement(item: unknown): item is BodyElement {
    return reflection.isInstance(item, BodyElement.$type);
}

export type BoolLit = 'false' | 'true';

export function isBoolLit(item: unknown): item is BoolLit {
    return item === 'true' || item === 'false';
}

export interface Button extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Button';
    actionRef?: string;
    caption: string;
    mods: Array<LayoutMod>;
}

export const Button = {
    $type: 'Button',
    actionRef: 'actionRef',
    caption: 'caption',
    mods: 'mods'
} as const;

export function isButton(item: unknown): item is Button {
    return reflection.isInstance(item, Button.$type);
}

export interface Card extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Card';
    body: Array<BodyElement>;
    caption?: string;
    mods: Array<LayoutMod>;
}

export const Card = {
    $type: 'Card',
    body: 'body',
    caption: 'caption',
    mods: 'mods'
} as const;
// ...
export interface Column extends langium.AstNode {
    readonly $container: Table;
    readonly $type: 'Column';
    caption: string;
    field: FieldRef;
    mods: Array<ColumnModifier>;
}

export const Column = {
    $type: 'Column',
    caption: 'caption',
    field: 'field',
    mods: 'mods'
} as const;
// ...
export type ColumnModifier = ColumnFormat | ColumnWidth;

export const ColumnModifier = {
    $type: 'ColumnModifier'
} as const;
// ...
export interface Confirm extends langium.AstNode {
    readonly $container: ToolbarNew | ViewAction;
    readonly $type: 'Confirm';
    message: string;
}

export const Confirm = {
    $type: 'Confirm',
    message: 'message'
} as const;

export function isConfirm(item: unknown): item is Confirm {
    return reflection.isInstance(item, Confirm.$type);
}

export interface Divider extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Divider';
}

export const Divider = {
    $type: 'Divider'
} as const;
// ...
export interface FieldRef extends langium.AstNode {
    readonly $container: Bind | Column | Table | Widget;
    readonly $type: 'FieldRef';
    domain: IdentName;
    field: IdentName;
}

export const FieldRef = {
    $type: 'FieldRef',
    domain: 'domain',
    field: 'field'
} as const;
// ...
export interface Grid extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Grid';
    body: Array<BodyElement>;
    cols: number;
    mods: Array<LayoutMod>;
}

export const Grid = {
    $type: 'Grid',
    body: 'body',
    cols: 'cols',
    mods: 'mods'
} as const;

export function isGrid(item: unknown): item is Grid {
    return reflection.isInstance(item, Grid.$type);
}

export type IconModifier = IconSize | LayoutMod;

export const IconModifier = {
    $type: 'IconModifier'
} as const;
// ...
export interface IconWidget extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'IconWidget';
    mods: Array<IconModifier>;
    name: string;
}

export const IconWidget = {
    $type: 'IconWidget',
    mods: 'mods',
    name: 'name'
} as const;
// ...
export type LayoutMod = Cols | Expand | Gap | Margin | Padding | ScrollFlag;

export const LayoutMod = {
    $type: 'LayoutMod'
} as const;

export function isLayoutMod(item: unknown): item is LayoutMod {
    return reflection.isInstance(item, LayoutMod.$type);
}

export interface LinkWidget extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'LinkWidget';
    caption: string;
    href: string;
    mods: Array<LayoutMod>;
}

export const LinkWidget = {
    $type: 'LinkWidget',
    caption: 'caption',
    href: 'href',
    mods: 'mods'
} as const;
// ...
export type NavModifier = As | Bind | Confirm;

export const NavModifier = {
    $type: 'NavModifier'
} as const;
// ...
export interface RichText extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'RichText';
    spans: Array<RichTextSpan>;
}

export const RichText = {
    $type: 'RichText',
    spans: 'spans'
} as const;

export function isRichText(item: unknown): item is RichText {
    return reflection.isInstance(item, RichText.$type);
}

export interface RichTextSpan extends langium.AstNode {
    readonly $container: RichText;
    readonly $type: 'RichTextSpan';
    style: RichTextStyle;
    text: string;
}

export const RichTextSpan = {
    $type: 'RichTextSpan',
    style: 'style',
    text: 'text'
} as const;
// ...
export interface Row extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Row';
    body: Array<BodyElement>;
    mods: Array<LayoutMod>;
}

export const Row = {
    $type: 'Row',
    body: 'body',
    mods: 'mods'
} as const;
// ...
export interface Section extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Section';
    body: Array<BodyElement>;
    caption?: string;
    mods: Array<LayoutMod>;
}

export const Section = {
    $type: 'Section',
    body: 'body',
    caption: 'caption',
    mods: 'mods'
} as const;

export function isSection(item: unknown): item is Section {
    return reflection.isInstance(item, Section.$type);
}

export interface Separator extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Separator';
}

export const Separator = {
    $type: 'Separator'
} as const;
// ...
export interface Stack extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Stack';
    body: Array<BodyElement>;
    mods: Array<LayoutMod>;
}

export const Stack = {
    $type: 'Stack',
    body: 'body',
    mods: 'mods'
} as const;
// ...
export interface Tab extends langium.AstNode {
    readonly $container: Tabs;
    readonly $type: 'Tab';
    body: Array<BodyElement>;
    caption: string;
}

export const Tab = {
    $type: 'Tab',
    body: 'body',
    caption: 'caption'
} as const;

export function isTab(item: unknown): item is Tab {
    return reflection.isInstance(item, Tab.$type);
}

export interface Table extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Table';
    actions: Array<ViewAction>;
    columns: Array<Column>;
    rowSource: FieldRef;
}

export const Table = {
    $type: 'Table',
    actions: 'actions',
    columns: 'columns',
    rowSource: 'rowSource'
} as const;

export function isTable(item: unknown): item is Table {
    return reflection.isInstance(item, Table.$type);
}

export interface Tabs extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Tabs';
    mods: Array<LayoutMod>;
    tabs: Array<Tab>;
}

export const Tabs = {
    $type: 'Tabs',
    mods: 'mods',
    tabs: 'tabs'
} as const;

export function isTabs(item: unknown): item is Tabs {
    return reflection.isInstance(item, Tabs.$type);
}

export interface ToolbarDelete extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'ToolbarDelete';
    confirm?: string;
}

export const ToolbarDelete = {
    $type: 'ToolbarDelete',
    confirm: 'confirm'
} as const;

export function isToolbarDelete(item: unknown): item is ToolbarDelete {
    return reflection.isInstance(item, ToolbarDelete.$type);
}

export interface ToolbarExit extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'ToolbarExit';
}

export const ToolbarExit = {
    $type: 'ToolbarExit'
} as const;

export function isToolbarExit(item: unknown): item is ToolbarExit {
    return reflection.isInstance(item, ToolbarExit.$type);
}

export type ToolbarItem = ToolbarDelete | ToolbarExit | ToolbarNew | ToolbarRefresh | ToolbarSave;

export const ToolbarItem = {
    $type: 'ToolbarItem'
} as const;

export function isToolbarItem(item: unknown): item is ToolbarItem {
    return reflection.isInstance(item, ToolbarItem.$type);
}

export interface ToolbarNew extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'ToolbarNew';
    caption: string;
    navMods: Array<NavModifier>;
    target: string;
}

export const ToolbarNew = {
    $type: 'ToolbarNew',
    caption: 'caption',
    navMods: 'navMods',
    target: 'target'
} as const;

export function isToolbarNew(item: unknown): item is ToolbarNew {
    return reflection.isInstance(item, ToolbarNew.$type);
}

export interface ToolbarRefresh extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'ToolbarRefresh';
    caption?: string;
}

export const ToolbarRefresh = {
    $type: 'ToolbarRefresh',
    caption: 'caption'
} as const;

export function isToolbarRefresh(item: unknown): item is ToolbarRefresh {
    return reflection.isInstance(item, ToolbarRefresh.$type);
}

export interface ToolbarSave extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'ToolbarSave';
}

export const ToolbarSave = {
    $type: 'ToolbarSave'
} as const;

export function isToolbarSave(item: unknown): item is ToolbarSave {
    return reflection.isInstance(item, ToolbarSave.$type);
}

export interface ViewAction extends langium.AstNode {
    readonly $container: Table;
    readonly $type: 'ViewAction';
    event: ViewActionEvent;
    navMods: Array<NavModifier>;
    target: string;
}

export const ViewAction = {
    $type: 'ViewAction',
    event: 'event',
    navMods: 'navMods',
    target: 'target'
} as const;

export function isViewAction(item: unknown): item is ViewAction {
    return reflection.isInstance(item, ViewAction.$type);
}

export type ViewActionEvent = 'on_edit' | 'on_new' | 'on_select';

export function isViewActionEvent(item: unknown): item is ViewActionEvent {
    return item === 'on_select' || item === 'on_new' || item === 'on_edit';
}

export interface ViewDecl extends langium.AstNode {
    readonly $container: ViewModel;
    readonly $type: 'ViewDecl';
    autoRefresh?: AutoRefresh;
    body: Array<BodyElement>;
    default: boolean;
    domains: Array<DomainBinding>;
    name: string;
    title: string;
    toolbar: Array<ToolbarItem>;
}

export const ViewDecl = {
    $type: 'ViewDecl',
    autoRefresh: 'autoRefresh',
    body: 'body',
    default: 'default',
    domains: 'domains',
    name: 'name',
    title: 'title',
    toolbar: 'toolbar'
} as const;
// ...
export interface Widget extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Widget';
    bind: FieldRef;
    caption?: string;
    kind: WidgetKind;
    mods: Array<WidgetModifier>;
}

export const Widget = {
    $type: 'Widget',
    bind: 'bind',
    caption: 'caption',
    kind: 'kind',
    mods: 'mods'
} as const;

export function isWidget(item: unknown): item is Widget {
    return reflection.isInstance(item, Widget.$type);
}

export interface WidgetExpand extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetExpand';
}

export const WidgetExpand = {
    $type: 'WidgetExpand'
} as const;

export function isWidgetExpand(item: unknown): item is WidgetExpand {
    return reflection.isInstance(item, WidgetExpand.$type);
}

export interface WidgetFocus extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetFocus';
}

export const WidgetFocus = {
    $type: 'WidgetFocus'
} as const;

export function isWidgetFocus(item: unknown): item is WidgetFocus {
    return reflection.isInstance(item, WidgetFocus.$type);
}

export interface WidgetFormat extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetFormat';
    fmt: FormatToken;
}

export const WidgetFormat = {
    $type: 'WidgetFormat',
    fmt: 'fmt'
} as const;
// ...
export interface WidgetMax extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetMax';
    value: number;
}

export const WidgetMax = {
    $type: 'WidgetMax',
    value: 'value'
} as const;

export function isWidgetMax(item: unknown): item is WidgetMax {
    return reflection.isInstance(item, WidgetMax.$type);
}

export interface WidgetMin extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetMin';
    value: number;
}

export const WidgetMin = {
    $type: 'WidgetMin',
    value: 'value'
} as const;

export function isWidgetMin(item: unknown): item is WidgetMin {
    return reflection.isInstance(item, WidgetMin.$type);
}

export type WidgetModifier = LayoutMod | WidgetExpand | WidgetFocus | WidgetFormat | WidgetMax | WidgetMin | WidgetPlaceholder | WidgetReadOnly | WidgetStep;

export const WidgetModifier = {
    $type: 'WidgetModifier'
} as const;

export function isWidgetModifier(item: unknown): item is WidgetModifier {
    return reflection.isInstance(item, WidgetModifier.$type);
}

export interface WidgetPlaceholder extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetPlaceholder';
    value: string;
}

export const WidgetPlaceholder = {
    $type: 'WidgetPlaceholder',
    value: 'value'
} as const;

export function isWidgetPlaceholder(item: unknown): item is WidgetPlaceholder {
    return reflection.isInstance(item, WidgetPlaceholder.$type);
}

export interface WidgetReadOnly extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetReadOnly';
}

export const WidgetReadOnly = {
    $type: 'WidgetReadOnly'
} as const;

export function isWidgetReadOnly(item: unknown): item is WidgetReadOnly {
    return reflection.isInstance(item, WidgetReadOnly.$type);
}

export interface WidgetStep extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetStep';
    value: number;
}

export const WidgetStep = {
    $type: 'WidgetStep',
    value: 'value'
} as const;